Write a get_int function (start_message, error_message, end_message) that takes three lines as arguments.
The function should ask the user for input until an integer is entered (a string accepted by the int function without errors).

Before the first input request, the start_message argument should be displayed, after each erroneous input, the value of the error_message string should be displayed and, if successful, the end_message string should be output and the resulting integer should be returned from the function (see the operation example). Each message displayed should be on a separate line.